// src/pages/project/model/types.ts
export type FormState = {
  manager_name: string;
  number_of_our_project: string;
  customer: string;
  name_of_project: string;
  type: string;
  po_number: string;
  customer_project_number: string;
  contact_person: string;
  delivery_date: string;
  sales_price: string;
  initial_cost: string;
  delivery_method: string;
  delivery_by_nmh: boolean;
};